Use relative URLs for submodules
Currently, committed URLs are HTTPS.
This requires the use of a token and frequent use
of credentials.
SSH URLs on the other hand allows a more convenient workflow.

We could convert all URLs to ssh instead, but then it becomes
inconsistent if someone clones SolAR in HTTPS (e.g. for read only
purpose), and don't want to configure SSH keys.

The solution is to use relative path, as described here:

https://www.damirscorner.com/blog/posts/20210423-ChangingUrlsOfGitSubmodules.html

This way, the submodule will be fetched with whatever method used
to clone SolAR.

This may lead to ssh errors when 'git submodule foreach' is used with
commands demanding a connection to the server (fetch, pull, ...).

A solution is to increase 'MaxStartups' in sshd config as described here:
https://docs.gitlab.com/ee/topics/git/troubleshooting_git.html#ssh_exchange_identification-error